Add files via upload
The primary goal of this project was to develop a predictive model capable of forecasting the delivery time of food orders in real-time. By analyzing various characteristics such as distance between the restaurant and the delivery location, as well as historical delivery times for similar distances, the intention was to create a tool that could assist delivery partners and restaurant owners in optimizing their operations and improving customer satisfaction.